requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.1.0
python-dotenv>=1.0.0
plotly>=5.17.0
streamlit-aggrid>=0.3.4
streamlit-option-menu>=0.3.6
streamlit-lottie>=0.0.5

